Support style hierarchy and more style properties
A new class StyleManager stores a full style hierarchy. It provides
methods to query a style property from a specific style element. The
style hierarchy will be traversed from the style element up to its
parent styles. A normal property will return the first value found. A
factor property will multiply with a factor-suffixed modifier style, if
any. A color property will incorporate the opacity style, if any, and
will blend with a blend-suffixed modifier style, if any.

BaseDataProviderTimeGraphProvider creates a new StyleManager instance
every time the list of style changes. It create the table of state items
used by the legend using the style manager. It updates the style
properties from the user preferences stored by the legend. It returns
the element style for a time event from the state model. It provides
interfaces to query the style hierarchy for an element style.

The interface IStylePresentationProvider is added to be implemented by
presentation providers that support styles.

TimeGraphControl will use this interface to query the style properties
for presentation providers that support it. In this case, this replaces
the use of the state table and state table index.

For drawing markers, the supported properties are: COLOR, OPACITY,
HEIGHT, SYMBOL_TYPE.

For drawing arrows, the supported properties are: COLOR, OPACITY,
HEIGHT.

For drawing states, the supported properties are: BACKGROUND_COLOR,
OPACITY, HEIGHT, BORDER_STYLE, BORDER_COLOR, BORDER_WIDTH.

The border is now drawn after the background is filled.

The use of the color registry is consolidated in a method.

[Added] Support for style hierarchy
[Added] Support more style properties in time graph control

Change-Id: I475a9443b3779cd321c30718cb26d614ed69903c
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/149389
Tested-by: CI Bot
Reviewed-by: Simon Delisle <simon.delisle@ericsson.com>
Tested-by: Simon Delisle <simon.delisle@ericsson.com>