Support scopes and id table in tree data providers
A scope is a unique string that can link the entry ids from different
data providers. If the scope is non-null, all data providers that share
the same scope have a common set of entry ids. If the scope is null, the
scope is unique to this single data provider.

An id table is a map that associates a key with an entry id. When used
in conjunction with a scope, this allows for multiple data providers to
share a unique entry id identified with a specific key.

[Added] Support for scopes and id table in tree data providers

Change-Id: I3ef6a99faa89ce0c5704143349dd30f09724cae6
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/150023
Tested-by: CI Bot
Tested-by: Simon Delisle <simon.delisle@ericsson.com>
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Simon Delisle <simon.delisle@ericsson.com>