tmf: Remove full list optimization in AbstractTimeGraphView zoom thread
When zooming out to the full time range, AbstractTimeGraphView skips
computation of the zoomed event lists and sets them to null, to fall
back on the full event list (normally computed once in build thread).

This is an unnecessary optimization. The zoomed event lists would have
to be computed if the zoom range was set to 99% of the full range, so
there is only a limited benefit to having this special case for 100%.

Furthermore, some implementations may be negatively affected by having
the base class unexpectedly skipping the normal zoom thread execution.

Change-Id: I3418ecc14da187e57b2ef65dacc097608ffe558c
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/150745
Tested-by: CI Bot
Reviewed-by: Simon Delisle <simon.delisle@ericsson.com>
Tested-by: Simon Delisle <simon.delisle@ericsson.com>