Support tooltip information from multiple data providers
BaseDataProviderTimeGraphPresentationProvider does not know from which
data provider tooltip information should be fetched. It now receives a
tooltip resolver for each data provider created by the
BaseDataProviderTimeGraphView, which can translate from (time event,
data provider) to the correct entry id and query the data
provider, only if it is a provider of that time graph entry. The results
from each query are aggregated, and some data providers can return no
tooltip information.

The model output element (either state, arrow or annotation) is passed
as a new fetch parameter to the tooltip query, so the data provider
knows which element is being queried.

In order to achieve this, a new parent model element ITimeElement is
created, with extending interfaces ITimeGraphState, ITimeGraphArrow and
IAnnotation. The model field of TimeEvent is changed to this parent
interface, since TimeEvent is extended with TimeLinkEvent
(ITimeGraphArrow) and MarkerEvent (IAnnotation).

ITimeElement extends a new interface IOutputElement, which represents
graphical elements that have a style (or a value) and support
element-based filtering.

The class OutputElement is created as base class of TimeGraphState,
TimeGraphArrow and Annotation.

[Added] Support tooltip information from multiple data providers

Change-Id: Ibbd1a11325d16f3875ce3430ce4ac38ea115b9f7
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/150807
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>