tmf: Use a code with the status of openTraceFromPath
When the status is OK after running the openTraceFromPath method, there
are 2 possibilities, either the trace will eventually be opened or the
trace type could not be set, the file is imported, but the trace won't
open. To differentiate those 2 cases (necessary when opening from the
command line, as we wait for the trace opened signal to know the trace
is fully opened), a code is added to the returned status.

OK: means the trace will eventually open

INFO: means the trace type could not be set (or any other eventual
information)

Change-Id: Ic78cf4271d411f41e3711cb8ba3cc3056c27c7e6
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/153548
Tested-by: CI Bot
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>