Fix BookmarksViewTest
When the editor is opened by double-clicking a bookmark, a blank table
is initially created and then the editor is reused, disposing the old
table and creating a new one.

The test was getting the table bot only once. It might be the empty
table that is later disposed. It should try to get the latest table from
the editor at every iteration of the wait condition.

Change-Id: I62c096a6c6e5b0522d8a40197e0411f23f66aa35
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/155926
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>