logging: Update log utils classes to match the spec
The TraceCompassLogUtils classes log statements according to the
trace-event specification. The flow events 's' and 'f' are not supposed
to create stack element, but they should be enclosed in a scope and be
represented only as arrows.

This patch replaces the 's' and 'f' beginning and end log entries of the
FlowScopeLog by Duration events 'B' and 'E' and adds an entry 's' or 't'
for an arrow endpoint.

Also adds the pid, which would be the same as tid here, since there is
no access to actual pid. Chromium trace-event reader requires the pid
field.

Also prints the timestamp as micro-second, as the trace-event spec
stipulates.

[changed] TraceCompassLogUtils implements the complete trace-event specification

Change-Id: I92bb6d8a9dab0ee6a046f153196d587d36ce39d3
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/155943
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>