os: Let ThreadStatusDataProvider return TimeGraphEntryModels
The ThreadEntryModel object should be only for entries that really
represent threads, the trace entry can be a simple TimeGraphEntryModel.
This requires the data provider to return trees of type
TimeGraphEntryModel instead of ThreadEntryModel, but if the model is a
ThreadEntryModel, then there is no special case of return value to
ignore.

Fixes bug 558208

[changed] ThreadStatusDataProvider return TimeGraphEntryModel elements
[fixed] Trace entry in Control Flow does not have a tid/pid

Change-Id: I133cceedfc64b09334dc0aa04eadda856fc7d345
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/156547
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>