io: Fix the request merges and inserts in state system
Many events can insert requests in queue for the same sector or they can
be of different types, so instead of a map with sector as key, queues are
a list. Also when issuing a request, sometimes there are other requests
covered by the same issued request and they need to be removed from the
waiting queue as well.

[Fixed] Disk request queues are cleaned up as requests for same sector are completed/moved

Change-Id: I53de9a65a660d85adeff3394f0e2b66553c1dbe4
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/157233
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>