tmf: Fix progress reporting of analyses
TmfAbstractAnalysis begins a task with UNKNOWN total work on the
received monitor. This prevents the progress reported by the analysis
execution to be displayed properly.

A SubMonitor is now used instead and passed to the analysis execution.

The call to done() is also removed, it is not the responsibility of the
callee to do so.

[Fixed] Fix progress reporting of analyses

Change-Id: I615779358663f521ea9aec9df6a792801d1813f4
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/157445
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>