Bug 560477: state system: Serialize unknown state types as strings
If a state object is not one of the known object types (Null, Long, Int,
Double, String, or CustomStateValue) then an IllegalArgumentException
was thrown. For example, when using an XML analysis with a CTF trace
which has an event field of type CtfEnumPair and this CtfEnumPair is
used as state value of an attribute, then such an exception would be
created.

This patch will convert the state value to a string, if the value object
is not one of the known types.

[Fixed] Bug 560477: IllegalArgumentException in SS serialization

Change-Id: I96f3b7cc3df7113579531251a98235a219a4d0c0
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/158248
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>