ctf: Support CTF enums as bit flags
bug 560479

When CTF enum values is not found in the enumeration, we try to see if
it can be a bit flag enum by trying to find a value for each bit that is
set in the value. If each bit has a value that does not correspond to a
range, then the result is a ORed string of all the bit values.

[added] Support of CTF enums as bit flags fields

Change-Id: I09a8e2ab027a6b4cde8210958536cb37987168ea
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/157665
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>