tmf: Fix handling of default styles
Return the factor modifier (if any) instead of null in
StyleManager.getFactorStyle() if there is no factor property set.

Return null instead of (0, 0, 0, 255) in StyleManager.getColorStyle() if
there is neither a color property nor an opacity modifier set.

In BaseDataProviderTimeGraphPresentationProvider.updateStyles(), do not
update the color and height properties to the default if they are null
in the base style and the legend setting matches the default.

Change-Id: Ie183c52bbcdc61eaae08376aa9dcfdb1540d316e
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/158614
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>