XY model: Return series data Collection instead of Map
XY data provider API return a Map of String to Model, but the outputs
should rather use the ID, which is common between tree and xy data
instead of a String, which is often an additional data that needs to be
matched for a given ID, and moreover is required to be unique per entry.

The API now returns only the model elements, which include the ID and
the name is not required anymore.

[Deprecated] ITmfXyModel#getData is replaced with ITmfXyModel#getSeriesData
[Deprecated] ITmfCommonXAxisModel#getYData is replaced with ITmfCommonXAxisModel#getYSeriesData
[Deprecated] AbstractTreeCommonXDataProvider#getYModels is replaced with AbstractTreeCommonXDataProvider#getYSeriesModel

Change-Id: Ibad424932a7b2279afbccffed41c77067bdb0421
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/159344
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>