tmf.ui: Add a ITmfFilterableControl interface for filters with '/'
For now, only the time graph views could be filtered locally with '/',
now with this interface, the dialog can be re-used with all views and
viewers that implement this interface.

If the view is in the context of
org.eclipse.tracecompass.tmf.ui.view.context,
then the '/' key will handle the interface calls.

[Added] ITmfFilterableControl interface for filters with the '/' key

Change-Id: I69df83e43247971b22cb0cc7a668143846610ae1
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/160439
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>