swtbot: Fix TimeGraphViewTest.testHideEmptyRows()
When filter is applied, multiple consecutive zoom threads can be started
as new entries become visible due to other entries being hidden. In
between these zoom threads there might be short periods where the time
graph is not considered dirty, which makes the unit test continue before
the view is completely updated with the filter results.

Replace use of the isDirty() condition with a condition that checks for
the expected number of visible items.

Change-Id: I3ad480b79c74fd97d5f3c1e2e2b850176e06d74c
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/161288
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>