os: bug 562415 Fix Control Flow view filtering
In patch 53b2d845933de51f6b2846c98c604ff95bbf1c80, the filterData method
of the thread status data provider was updated and was using the wrong
parameter key to send the entry ID, which caused the tooltips to return
nothing.

Also, the fetchTooltip method is updated to use the fetch parameter map
directly instead of converting to query and uses a 2D query to get the 2
necessary states instead of full query.

Change-Id: I2c7b169b0685d872547b9ef22e9293fb102dfcaf
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/161407
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>