tmf.ui: Add instance ID to AbstractSelectTreeViewer
AbstractSelectTreeViewer uses class' name as part of key for
trace context data. There are cases when it might be needed
to distinguish contexts between instances of the same class.
This change adds an ability to set instance's arbitrary ID
that is aimed to distinguish this instance's data from others
of the same class.

Change-Id: I99223d0ab0f1c93dd6d68cd89162e8315c6083df
Signed-off-by: Ivan Grinenko <ivan.grinenko@auriga.com>
Reviewed-on: https://git.eclipse.org/r/163745
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>