tmf: Fix initialization of TmfFilteredXYChartViewer
The method initializeDataProvider() is called every time there is a UI
refresh. It should not add the data provider to the presentation
provider every time, as its provider list can contain duplicates.

The flag fUseDefaultStyleValues is global and never reset. Once set to
false for one trace it remains false for all other traces. The flag
should be stored in a map so that each value is independent.

The creation of the presentation provider in TmfCommonXAxisChartViewer
should not be called every time a trace becomes active when switching
back and forth opened traces. It should only be created once.

[Fixed] tmf: Fix initialization of TmfFilteredXYChartViewer

Change-Id: I56b3d1cd40b31b97a8104142f0f57790c6bd3230
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/163954
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>