os: bug 563828 Set total series earlier in tree viewers
The entries representing total and shown with different styles are known
when they are received. They should be set at that moment in the
presentation provider to avoid race condition when the style is fetched
by the XY chart before the entry ID has been marked as a 'total' series
by the tree.

[Fixed] Fix a race condition when getting the style of XY charts when total series are not set yet by the tree.

Change-Id: Iaeca6f89eaee2d2773893edcd973121e4b41c9c4
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/164000
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>