tmf: Fix series width in XYPresentationProvider
Revert the change from https://git.eclipse.org/r/163935.

The value of StyleProperties.WIDTH should be an integer after all. The
Javadoc is updated to show the default value as 1.

The method StyleManager.getFactorStyle() is updated to accept any type
of Number as the base value. The method will still always return a
Float, and users can convert it back to an integer if necessary.

Change-Id: Ie48d78071095092bfb88193c4c348c47d39f71bb
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/164025
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>