tmf: Support WIDTH property for arrow style
Instead of using HEIGHT style property with a default value of 0.1f, use
the WIDTH property with a default value of 1 (Integer).

Use the integer width to set the arrow line width if available,
otherwise use the height for backward compatibility.

Update the legend to use a scale between 1 and 10 when the style map
contains the WIDTH key.

Store the state width as an integer in the preference store.

[Added] Support WIDTH property for arrow style

Change-Id: I0f6f44aeded9b24900c46c7cd9024356a87036ca
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/164945
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>