tmf: Support COLOR style property in TimeGraphLegend
The COLOR property should be used for arrows and annotations. The
BACKGROUND_COLOR property should be used for states.

The BaseDataProviderTimeGraphPresentationProvider only supported
BACKGROUND_COLOR when creating the style map for the StateItem used by
the TimeGraphLegend to initialize its color widgets, so the data
provider implementations had to set both COLOR and BACKGROUND_COLOR for
arrows and annotations. The legend would initialize itself with the
BACKGROUND_COLOR but would change both properties when changed by user.

The BaseDataProviderTimeGraphPresentationProvider now sets both
BACKGROUND_COLOR and COLOR independently in the StateItem's style map
when these properties are set in the data provider's style map. It no
longer forces a default BACKGROUND_COLOR if none is set.

The legend will initialize itself with the StateItem's style map using
the BACKGROUND_COLOR, or the COLOR, or the default color, whichever one
is found first in that order.

[Fixed] Support COLOR style property in TimeGraphLegend

Change-Id: Iec7f8c72005e2c673297cf9d62c01140e95aacd6
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/165190
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>