tmf: Fix handling of HEIGHT for symbol markers
The HEIGHT style property is defined so that height value 1.0f means the
normal element height. For a symbol marker this should be implemented so
that the normal element height is equal to the height of the time graph
entry, or to the default of 4 pixels for a XY chart.

The symbolSize parameter in the SymbolHelper methods is changed to mean
half of the height of the symbol. This is the same as the radius for
symmetrical symbols, and is more clearly defined for asymmetrical ones.

The triangle symbols are changed so that they draw to the full height on
both sides.

These last two changes also make SymbolHelper match the drawing
implementation in org.eclipse.swtchart plug-in.

[Fixed] Fix handling of HEIGHT for symbol markers

Change-Id: Ia249c60b251e324dff1f33977cbd727897a3e8d8
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/165494
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>