Revert "analysis: Add cell descriptors for data type/unit to ITmfTreeDataModel"
Instead of describing the type of the value and having each client
implement the formatting of the strings for each data type and unit,
let the data providers provide the proper formatted string. The data
providers have the domain specific knowledge to do the proper
formatting, and this will simplify the client (UI) implementation.

[Deleted] table cell descriptors in ITmfTreeDataModel (reverted)

This reverts commit 5584c91b2112d172be95222398753bcc94b00e45.

Change-Id: I53fd092a175d1cf7f927f28187956cccb79b217f
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/166760
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>