tmf: Add a preference to not hide empty rows on filter
When applying a filter in a timegraph, rows that do not have any
filtered items in the time range are hidden. This behavior is already
covered by the "Hide Empty Rows" button, and it should be the user's
choice to enable or disable it.

Since many users are already used to the previous behavior, the user can
choose between automatically hiding rows upon filtering or not with a
preference under the "Tracing" category

[Fixed] Bug 565097: When filtering, Time graph rows disappear when empty
[Added] Preference option for hiding empty rows upon filtering

Change-Id: I3df84076a77f6ca1cc54b559805395d9c5a0d349
Signed-off-by: David Piché <dpiche.veg@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/167031
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>