root: add logging.properties
A sample logger to allow for easy tracing of trace compass.

to use:

Launch trace compass or the trace server or whatever derived
product with the following lines

(java 8 only) -vmargs
-Djava.util.logging.config.file=%gitroot%/logging.properties
-Dorg.eclipse.tracecompass.logging=true

the output json needs to replace every } "" { with }, {
and the first character needs to be a [,
the last character needs to be a ]

Using the jsonify.sh script if possible to convert the output
to the trace-event format.

See README.md or Developper guide for more information.

[Added] Sample logging configuration

Change-Id: If727f3404fd52725fdff7832f66fa19803e07f7d
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/167029
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>