tmf: Fix ResourceUtilTest
It appears that the @Parameters method can be called twice in newer
version of Maven, leading to a ResourceException.

It is also called before running all the tests, which means that the
code in @Parameters could interfere with other tests if it modifies the
workspace.

Move the creation of projects to the @BeforeClass method.

Change-Id: Ie8f4c22633aa04da230cb0e7a9921b401b1c4a1f
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/167126
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>