releng: Add e4.18 and update eStaging targets for 2020-12 M3
Fixes compilation errors due to the usage of deprecated APIs by
switching the settings to warning to be able to build for older
targets as well.

Addresses build failure described in Bug 569183 by commenting out the
relevant lines of code. Note that the performance reporting will be
fixed in a separate patch.

Eclipse CDT changed their dependency to jakarta.xml.bind instead of
javax.xml.bind. jakarta.xml.bind is only available on Orbit for 4.18
simultaneous release. Update all targets and the RCP target to include
both. With this it will be possible to build for older and new targets.

Change-Id: I1bcdf0301a831c70b396ec1d0d1b07722d225dfe
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/171771
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>