linux: Make Data Provider Factories work with TmfExperiment
The ThreadStatusDataProvider and ResourcesStatusProvider factories
didn't create a data provider if the passed ITmfTrace was and
experiment. When used in the BaseDataProviderTimeGraphView it's ok,
because the experiment handling is done in the view class itself.
However, when it's used differently (e.g. using the trace server,
then it' doesn't work and only the first found kernel trace is used
to create a data provider which is then returned. Because of that,
experiments with multiple kernel traces or experiments with UST traces
(for UST markers) were not working properly.

[Fixed] Resources Status and Thread Status DP factories for experiments
[Added] static create() method in TmfTimeGraphCompositeDataProvider

Change-Id: I52ff4ab2c9f3a13a34646274e29c1fad51eb597f
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/174641
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>