profiling: Fix callstack benchmarks (bis)
In august 2018, the callstack benchmark was fixed by using the relative
path to the downloaded benchmark traces.

In december 2019, a swtbot UI benchmark was added for one of the
benchmark traces, but for the test to work, it went back to using the
plugin's bundle path to get the trace. It broke the callstack benchmarks
again.

Now, we go back to using the relative path, but returned as absolute so
it works both for benchmarks and swtbot perf tests.

Change-Id: I4d45b082d495202adf8a323394777a4089df3e37
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/177263
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>