tmf.ui: throttle querying long running data providers
When the back-end is running, many resources are in use. So querying it
could slow down the act of indexing. Instead of querying every 500 ms,
query the first time 500 ms later, the next one takes 1s, 1.5s, etc...
This approach should keep responsiveness for short lived queries but not
overwhelm the back-end.

[Changed] Throttle back-end queries on trees and timegraphs

Change-Id: I6982d0c90a47e31755176a7be8f1cd61c445198d
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/177516
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>