kernel.tests: add a benchmark to test experiments
In order to test Trace Compass scalability with large
number of traces (experiments) with diffrent analysis modules, this
benchmark has been added. It uses the parent directory
path containg the traces in order to load them and run the test.
The default analysis module used in this benchmark is "KernelAnalysisModule"
but it can be replaced by the any other analysis.

Change-Id: I2efd2f098c71da3f6ec15650dd5f21e12d6bedd4
Signed-off-by: Abdellah Rahmani <rahmaniabdellah1994@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/178011
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>