kernel.test: add benchmark to test experiments indexing
In order to test Trace Compass scalability with large
experiments of different sizes, this benchmark has been
added. It takes the parent directory path containing the
Lttng Kernel traces as an input in order to load them
and run the test. The path is by default set to "null",
so it should be changed before running the test.

Change-Id: I382b47efd68e493a6f2d355495387d82325d488a
Signed-off-by: Abdellah Rahmani <rahmaniabdellah1994@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/178231
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>