kernel.test: add a bechmark to test Lttng Analyses
This benchmark uses PerformanceMeters to measure the CPU
and memeory usage for each trace for the chosen analysis.
It allows the user to load his traces directly from the
main directory making it easy when we have many traces to
test.The path to the traces directory should be then
provided by the user before running this benchmark.

Change-Id: I2d62e91b33e954ab5d7aa83e795704f26ac029d5
Signed-off-by: Abdellah Rahmani <rahmaniabdellah1994@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/178236
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>