ctf.test: add a bechmark to test trace reading with CTFTraceReader
This benchmark is added to be able to test Lttng traces reading with
the CTFTraceReader. The idea is to test traces of different sizes going
from some MBs to many GBs, this is why the user has to give a path
to the directory containing the traces before running this benchmark.

Change-Id: I9088efaee10f3778542763a9a914b07477325b58
Signed-off-by: Abdellah Rahmani <rahmaniabdellah1994@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/178266
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>