tmf.core: Add trace annotation provider and implement it for lost events
This annotation provider for global trace annotations, is meant to
replace the IMarkerEventSource base marker providers which have UI
dependencies. The new implementation can be in core only applications
like the Trace Compass trace server.

The IMarkerEventSource and it's implementations will be deprecated and
gradually migrated to the core trace annotation provider, as well as the
AbstractTimeGraphView will be updated to use the new trace annotation
provider.

[Added] TraceAnnotationProvider in core and implement it for lost events

Change-Id: I73df8da20ca6fb4ae44f09a3a21eccd6e474a681
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/178933
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>