ust.core: Fix realloc memory usage tracking
Current code tracks the realloc() memory usage using the argument
pointer. However, this pointer is not the one that should be tracked,
as the new allocated memory region is represented by the returned
pointer instead. As consequence of this behavior, the UST memory usage
chart could be incorrect in the presence of realloc events.

This commit change the described behavior to use the returned pointer
(ptr) instead of the argument pointer (inPtr).

[Fixed] LTTng ust tracks realloc properly.

Signed-off-by: João Petroni <jdspetroni@gmail.com>
Change-Id: I2a53978f266583c21a905b87212075d566e86838
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/178673
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>