tmf: Bug 567888: TmfTreeCompositeDataProvider doesn't consider scope
Allows the composite data provider to receive the same entryId from
multiple data providers. If they are within the same scope, a single
entry will be returned to the caller fetching the tree and duplicates
will be discarded.

The order of entries from a single data provider is preserved in most
cases but the order of independent entries from different data providers
is undefined.

[Fixed] Bug 567888: TmfTreeCompositeDataProvider doesn't consider scope
which can lead to ID clashes

Change-Id: I5aa416434623cadfaa295a813de8ef9da67c2e5d
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/178911
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>