tmf.core: Correctly hash long[]
Passing single null or primitive array argument to Object[] does not
work as expected. In the case of null, it is not passed as an array with
single element, but the argument itself is null. In the case of a
primitive array, if the formal parameter is Object...,
it is passed as a single element array.

Change-Id: I01ea599f6927e4ef0466ab062840076f894565c4
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/180731
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>